Fix #2, negative reading level
Now displays N/A instead of a negative number.
If one box is N/A, the average only displays the non-neg box. If both boxes are N/A, the average displays N/A.
Wrapped call to TextStatistics b/c it throws an error if trying to parse an empty string.